Move unwrapIncludeSrc into Page.js
